migrate tests into enchannel-zmq's main suite
